updated threading policy, again
X11 requires that all calls to the window API be done from the main thread,
unless you go through a little bit of extra effort with linking against X11
and calling XInitThreads. I didn't want to do that so I did this instead,
adding event queues for the UI.

I don't plan on having event queues for the game logic, instead I'll have the
ui directly call the game logic functions to update data. This way the UI
won't have to wait until the next gametick to be in sync with what's in the
game logic thread.