purchase buttons on the marketListingWidget
big commit!!! So it turns out initializing the widget with a vector of
pointers wasn't such a good idea. When a widget containing buttons is
copied, the addresses of the buttons it contains aren't updated in the
button list.

Solution is similar to "getSubWidgets", there's now a "getButtons" abstract
method for widgets.