overhaul how buttons work
jesus

how it worked before:
buttons exist as an ID. A screen checks if a mouse was pressed on a button
and notes its ID. When the mouse button is released the screen checks the
button IDs again, and if it matches the pressed one then that ID is sent to
a lookup of button functions.

how it works now:
buttons are just another widget. They monitor the mouse themselves and
check if the mouse was pressed and released on itself. A lambda function
is passed into the constructor which is what sends events off (if the button
was clicked).

This is part of a larger refactor to make everything widgets so the UI can
be rendered easier