property, market, and loan screen contain better widget heights
previously if there were 1 property on the market it would take up the whole
screen. It's now "fixed" to display 1/5 of the screen. Still dynamic and
changes actual pixel heights but a fixed percentage. In the future there'll
be some method of viewing more widgets than shown, maybe a scroll bar or
next button